- Moved ClientCommand related things to command.rs
- Fixed AOMessageCodec implementation

- Written some tests
