+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![forbid(unsafe_code)]
 
+pub mod command;
 pub mod config;
 pub mod master_server_client;
 pub mod networking;
